requirements.txt
annotated-types==0.6.0
greenlet==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
typing_extensions==4.10.0
